[geoclue-provider-hybris] Limit cellular technology when downloading AGPS data. Contributes to JB#18157
AGPS with SUPL v1 does not support LTE. When starting positioning limit
the allowed cellular technology to UMTS or lower. The original cellular
technology preference is saved and restored when no longer required by
AGPS. This is either when:

 * GPS stack indicates that data connection is no longer required
 * GPS stack returns position data
 * positioning stops

This temporarily changes a user visible setting. It is possible that
the user will notice and change it back in the settings application.
Resulting in failure to download AGPS data.