[geoclue-provider-hybris] Postpone removal of magnetic variation. Fixes JB#28443
The command that removes the magnetic variation dconf key needs to be
run as the user not root. Convert to a oneshot and schedule it to run
during the next boot.

Remove ldconfig from both post and postun sections. It does not need to
be run as there are no shared libraries in this package.