[geoclue-provider-hybris] Disable UTC time injection. Contributes to JB#28488
This is a work around for an assert triggered in gpsd on the Jolla
tablet. gpsd aborts with the message: 'Abnormal GlEngine stop. Initial
time estimate was bad'.

For now just disable time injection all together. This should not have
too much impact as the GPS on the Jolla phone reports that it does not
support GPS_CAPABILITY_ON_DEMAND_TIME and so never requests time
injection.