[geoclue-hybris] Fix SUPL assisted GPS. Fixes JB#31121
Asynchronous improvements to libqofono had the unfortunate effect of
breaking the code used to get the APN of the active cellular
connection. As an APN was never reported to the GPS, SUPL data was
never downloaded resulting to slow time to first lock.

Fixed by taking the asynchronous nature of the libqofono API into
account.