[geoclue-provider-hybris] Keep plugin active while position consumers exist. Contributes to JB#34741
Previously, the hybris plugin would only keep itself active if
gps positioning was enabled and it had position consumers.

This commit removes the first part of the requirement, so that the
plugin will now remain active (but disabled) if gps positioning
is disabled but it has position consumers.

For example, if a maps application is open, but the user has
disabled positioning via location.conf / settings, the plugin
will remain active but disabled, so that as soon as the user
enables positioning it can immediately provide position updates
to the position consumer (maps application).

Contributes to JB#34741